nothing of note

+++ src/test.cpp
@@ -15,7 +15,7 @@ int main(void){
 
     //inf_int<int8_t> a = 1000;
 
-    inf_int<int8_t> a = 200;
+    inf_int<int> a = 200;
 
     std::cout << a.get_buffer<int>() << N;
